requirements.txt
bcrypt==5.0.0
build==1.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
docutils==0.22.3
dotenv==0.9.9
id==1.5.0
idna==3.11
invoke==2.2.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
keyring==25.7.0
markdown-it-py==4.0.0
mdurl==0.1.2
more-itertools==10.8.0
nh3==0.3.2
packaging==25.0
paramiko==4.0.0
pycparser==2.23
Pygments==2.19.2
PyYAML==6.0.3
PyNaCl==1.6.1
pyproject_hooks==1.2.0
python-dotenv==1.2.1
pywin32-ctypes==0.2.3
readme_renderer==44.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.2.0
tqdm==4.67.1
twine==6.2.0
urllib3==2.6.1
